+++ src/extension.ts
@@ -47,8 +47,9 @@ export async function activate(context: vscode.ExtensionContext) {
     })
   );
   // Linter is always activated but will only lint if compiler !== Disabled
-  const linter = new FortranLintingProvider(logger);
-  linter.activate(context.subscriptions);
+  const linterCache = path.join(context.storageUri.fsPath);
+  const linter = new FortranLintingProvider(logger, linterCache);
+  context.subscriptions.push(...(await linter.activate()));
   vscode.languages.registerCodeActionsProvider(FortranDocumentSelector(), linter);
 
   if (formatterType !== 'Disabled') {