+++ src/features/commands.ts
@@ -4,6 +4,9 @@
 
 export const RestartLS = 'fortran.analysis.restartLanguageServer';
 export const RescanLint = 'fortran.analysis.rescanLinter';
+export const CleanLintDiagnostics = 'fortran.analysis.cleanLinterDiagnostics';
+export const CleanLintFiles = 'fortran.analysis.cleanLinterFiles';
+export const InitLint = 'fortran.analysis.initLinter';
 export const WhatsNew = 'fortran.analysis.showWhatsNew';
 export const BuildRun = 'fortran.build.run';
 export const BuildDebug = 'fortran.build.debug';